+++ .gitignore
@@ -5,6 +5,7 @@
 *.suo
 *.user
 *.sln.docstates
+*.jfm
 
 # Build results
 [Dd]ebug/
