+++ AgileSqlClub.SqlPackageFilter.IntegrationTests/FilterTests/XmlMultiFilterTests.cs
@@ -43,11 +43,11 @@ namespace AgileSqlClub.SqlPackageFilter.IntegrationTests
             sr.Dispose();
 
             var args = string.Format(
-                "/Action:Publish /TargetServerName:localhost /SourceFile:DacPac.dacpac /p:AdditionalDeploymentContributors=AgileSqlClub.DeploymentFilterContributor " +
+                $"/Action:Publish /TargetServerName:localhost /SourceFile:{Path.Combine(TestContext.CurrentContext.TestDirectory, "Dacpac.Dacpac")} /p:AdditionalDeploymentContributors=AgileSqlClub.DeploymentFilterContributor " +
                 " /TargetDatabaseName:Filters /p:DropObjectsNotInSource=True " +
                 "/p:AdditionalDeploymentContributorArguments=SqlPackageLogging=Info;\"SqlPackageXmlFilterFile={0}\" /p:AllowIncompatiblePlatform=true", xmlFile);
 
-            var proc = new ProcessGateway(".\\SqlPackage.exe\\SqlPackage.exe", args);
+            var proc = new ProcessGateway( Path.Combine(TestContext.CurrentContext.TestDirectory,   "SqlPackage.exe\\SqlPackage.exe"), args);
             proc.Run();
             proc.WasDeploySuccess();
