+++ AgileSqlClub.SqlPackageFilter.IntegrationTests/SqlPackage.exe/Extensions/Readme.txt
@@ -1,0 +1,1 @@
+http://go.microsoft.com/fwlink/?LinkID=287172
