+++ AgileSqlClub.SqlPackageFilter.IntegrationTests/SqlPackage.exe/Microsoft.SqlServer.DacAzureExtensions.xml
@@ -1,0 +1,77 @@
+<?xml version="1.0"?>
+<doc>
+    <assembly>
+        <name>Microsoft.SqlServer.DacAzureExtensions</name>
+    </assembly>
+    <members>
+        <member name="T:Microsoft.SqlServer.DacAzureExtensions.DacAzureResources">
+            <summary>
+              A strongly-typed resource class, for looking up localized strings, etc.
+            </summary>
+        </member>
+        <member name="P:Microsoft.SqlServer.DacAzureExtensions.DacAzureResources.ResourceManager">
+            <summary>
+              Returns the cached ResourceManager instance used by this class.
+            </summary>
+        </member>
+        <member name="P:Microsoft.SqlServer.DacAzureExtensions.DacAzureResources.Culture">
+            <summary>
+              Overrides the current thread's CurrentUICulture property for all
+              resource lookups using this strongly typed resource class.
+            </summary>
+        </member>
+        <member name="P:Microsoft.SqlServer.DacAzureExtensions.DacAzureResources.AkvFailedToObtainToken">
+            <summary>
+              Looks up a localized string similar to Failed to obtain the access token.
+            </summary>
+        </member>
+        <member name="P:Microsoft.SqlServer.DacAzureExtensions.DacAzureResources.AkvMissingPermissions">
+            <summary>
+              Looks up a localized string similar to Missing permissions for Azure Key Vault key &apos;{0}&apos;: {1}.
+            </summary>
+        </member>
+        <member name="P:Microsoft.SqlServer.DacAzureExtensions.DacAzureResources.AkvNoAccess">
+            <summary>
+              Looks up a localized string similar to Unable to access Azure Key Vault key &apos;{0}&apos;: {1}.
+            </summary>
+        </member>
+        <member name="P:Microsoft.SqlServer.DacAzureExtensions.DacAzureResources.AkvNoCredentials">
+            <summary>
+              Looks up a localized string similar to Azure Key Vault authentication settings were not configured. &apos;ClientId&apos; and &apos;Secret&apos; parameters must be specified as the current operation will result in data motion involving encrypted tables. .
+            </summary>
+        </member>
+        <member name="T:Microsoft.SqlServer.Dac.KeyVault.DacKeyVaultAuthenticator">
+            <summary>
+            Methods for configuring, verifying, and using Azure Key Vault with DacFx for Always Encrypted
+            </summary>
+        </member>
+        <member name="M:Microsoft.SqlServer.Dac.KeyVault.DacKeyVaultAuthenticator.GetToken(System.String,System.String,System.String)">
+            <summary>
+            <see cref="M:Microsoft.SqlServer.Dac.KeyVault.KeyVaultAuthenticator.GetToken(System.String,System.String,System.String)"/>
+            </summary>
+        </member>
+        <member name="M:Microsoft.SqlServer.Dac.KeyVault.DacKeyVaultAuthenticator.InitializeAuthenticator">
+            <summary>
+            <see cref="M:Microsoft.SqlServer.Dac.KeyVault.KeyVaultAuthenticator.InitializeAuthenticator"/>
+            </summary>
+        </member>
+        <member name="M:Microsoft.SqlServer.Dac.KeyVault.DacKeyVaultAuthenticator.UpdateAuthInfo(Microsoft.SqlServer.Dac.KeyVault.KeyVaultAuthInfoBase)">
+            <summary>
+            <see cref="M:Microsoft.SqlServer.Dac.KeyVault.KeyVaultAuthenticator.UpdateAuthInfo(Microsoft.SqlServer.Dac.KeyVault.KeyVaultAuthInfoBase)"/>
+            </summary>
+        </member>
+        <member name="M:Microsoft.SqlServer.Dac.KeyVault.DacKeyVaultAuthenticator.Validate(System.Collections.Generic.IList{System.String},System.Threading.CancellationToken)">
+            <summary>
+            <see cref="M:Microsoft.SqlServer.Dac.KeyVault.KeyVaultAuthenticator.Validate(System.Collections.Generic.IList{System.String},System.Threading.CancellationToken)"/>
+            </summary>
+        </member>
+        <member name="T:AssemblyRef">
+            <summary>
+            Sets public key string for friend assemblies.
+            </summary>
+        </member>
+        <member name="F:AssemblyRef.ProductPublicKey">
+            <summary>No signing</summary>
+        </member>
+    </members>
+</doc>
