+++ AgileSqlClub.SqlPackageFilter.IntegrationTests/SqlPackage.exe/sqlpackage.exe.config
@@ -8,7 +8,7 @@
       Uncomment the settings below to enable tracing.
       Update the initializeData for the TextLogger with the desired trace file path.
     -->
-
+    <!--
     <sources>
       <source name="Microsoft.Data.Tools.Diagnostics.Tracer" switchValue="Error">
         <listeners>
@@ -19,7 +19,7 @@
         </listeners>
       </source>
     </sources>
-
+    -->
   </system.diagnostics>
   <!--
   The below runtime setting is a workaround for an issue resulting from a known breaking change in the SQL Server 2012 version of the System CLR Types. The original breaking change and resulting issue are both documented in the “SQL CLR Data Types” section of http://go.microsoft.com/fwlink/?LinkID=271559.
@@ -28,7 +28,12 @@
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
       <dependentAssembly>
         <assemblyIdentity name="Microsoft.SqlServer.Types" publicKeyToken="89845dcd8080cc91" culture="neutral" />
-        <bindingRedirect oldVersion="10.0.0.0-11.0.0.0" newVersion="12.0.0.0" />
+        <bindingRedirect oldVersion="10.0.0.0" newVersion="13.0.0.0" />
+        <bindingRedirect oldVersion="12.0.0.0-13.0.0.0" newVersion="13.0.0.0" />
+      </dependentAssembly>
+      <dependentAssembly>
+        <assemblyIdentity name="Microsoft.SqlServer.TransactSql.ScriptDom" publicKeyToken="89845dcd8080cc91" culture="neutral" />
+        <bindingRedirect oldVersion="12.0.0.0-13.0.0.0" newVersion="13.100.0.0" />
       </dependentAssembly>
     </assemblyBinding>
   </runtime>
