+++ AgileSqlClub.SqlPackageFilter.IntegrationTests/app.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
 <configuration>
   <appSettings>
-    <add key="ConnectionString" value="SERVER=.;integrated security=sspi;initial catalog=filters"/> 
+    <add key="ConnectionString" value="SERVER=.;integrated security=sspi;initial catalog=Filters"/>
