+++ AgileSqlClub.SqlPackageFilter.UnitTests/packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Castle.Core" version="3.3.3" targetFramework="net451" />
+  <package id="Microsoft.SqlServer.DacFx.x64" version="130.3370.2" targetFramework="net451" />
   <package id="Moq" version="4.5.13" targetFramework="net451" />
   <package id="NUnit" version="3.4.1" targetFramework="net451" />
 </packages>
