+++ DacPac/Tables/Employees.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Employees]
 (
 	[EmployeeId] INT NOT NULL PRIMARY KEY,
-	[Name] VARCHAR(25) NOT NULL
+	[Name] VARCHAR(25) NOT NULL default  'aaaaa'
 )
