+++ SqlPackageFilter/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.SqlServer.DacFx.x64" version="130.3370.2" targetFramework="net40" />
+  <package id="Microsoft.SqlServer.DacFx.x64" version="130.3352.1" targetFramework="net40" />
 </packages>
